PAN-3 Update directory structure to a src/main src/test configuration
