+++ pandos/MANIFEST.in
