+++ pandos/src/main/pandos.egg-info/SOURCES.txt
@@ -1,0 +1,15 @@
+LICENSE
+MANIFEST.in
+README.md
+requirements.txt
+setup.py
+bin/pandos
+src/main/pandos/cli.py
+src/main/pandos/settings.py
+src/main/pandos/version
+src/main/pandos/version.py
+src/main/pandos.egg-info/PKG-INFO
+src/main/pandos.egg-info/SOURCES.txt
+src/main/pandos.egg-info/dependency_links.txt
+src/main/pandos.egg-info/requires.txt
+src/main/pandos.egg-info/top_level.txt
