+++ pandos/src/main/pandos.egg-info/requires.txt
@@ -1,0 +1,2 @@
+fire==0.5.0
+requests==2.31.0
