+++ pandos/src/main/pandos/settings.py
